Backend-Microplate-vision-inference-service/requirements.txt
fastapi==0.115.12
python-jose==3.4.0
ultralytics==8.3.127
opencv-python==4.11.0.86
python-dotenv==1.1.0
python-multipart==0.0.20
uvicorn==0.34.2
httpx==0.27.0
pydantic==2.5.0
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.0
scikit-learn==1.3.0
joblib==1.3.2
dill==0.4.0
pika==1.3.2